+++ package.json
@@ -3,7 +3,8 @@
   "packageManager": "pnpm@10.33.0",
   "scripts": {
     "build": "pnpm -r build",
-    "test": "pnpm -r --if-present test",
+    "clean": "pnpm -r --if-present clean",
+    "test": "pnpm build && pnpm -r --if-present test",
     "format": "biome format --write .",
     "check": "biome check --write .",
     "typecheck": "pnpm -r exec tsc --noEmit"
