+++ packages/dhcp/tsconfig.json
@@ -3,8 +3,10 @@
   "include": ["src/**/*.ts"],
   "compilerOptions": {
     "paths": {
-      "tcpip": ["../tcpip/src"],
-      "@tcpip/wire": ["../wire/src"]
+      "tcpip": ["../tcpip/src/index.js"],
+      "tcpip/types": ["../tcpip/src/types.js"],
+      "@tcpip/dns": ["../dns/src/index.js"],
+      "@tcpip/wire": ["../wire/src/index.js"]
     }
   }
 }
