+++ packages/tcpip/src/types.ts
@@ -1,48 +1,110 @@
-import type { CommonExports } from './bindings/base.js';
-import type {
-  BridgeExports,
-  BridgeInterface,
-} from './bindings/bridge-interface.js';
-import type {
-  LoopbackExports,
-  LoopbackInterface,
-} from './bindings/loopback-interface.js';
-import type { TapExports, TapInterface } from './bindings/tap-interface.js';
-import type { TcpExports } from './bindings/tcp.js';
-import type { TunExports, TunInterface } from './bindings/tun-interface.js';
-import type { UdpExports } from './bindings/udp.js';
-import type { UniquePointer } from './util.js';
+import type { IPv4Address, IPv4Cidr, MacAddress } from '@tcpip/wire';
 
-export type Pointer = UniquePointer | 0;
-
-export type WasiExports = {
-  memory: WebAssembly.Memory;
-  _initialize(): unknown;
+export type DuplexStream<R = unknown> = {
+  readable: ReadableStream<R>;
+  writable: WritableStream<R>;
 };
 
-export type SysExports = {
-  malloc(size: number): number;
-  free(ptr: number): void;
+export type UdpDatagram = {
+  host: string;
+  port: number;
+  data: Uint8Array;
 };
 
-export type StackExports = {
-  process_queued_packets(): void;
-  process_timeouts(): void;
+export type UdpSocketOptions = {
+  /**
+   * The local host to bind to.
+   *
+   * If not provided, the socket will bind to all available interfaces.
+   */
+  host?: string;
+  /**
+   * The local port to bind to.
+   *
+   * If not provided, the socket will bind to a random port.
+   */
+  port?: number;
 };
 
-export type WasmExports = WasiExports &
-  SysExports &
-  StackExports &
-  CommonExports &
-  LoopbackExports &
-  TunExports &
-  TapExports &
-  BridgeExports &
-  TcpExports &
-  UdpExports;
+export type UdpSocket = {
+  readable: ReadableStream<UdpDatagram>;
+  writable: WritableStream<UdpDatagram>;
+  close(): Promise<void>;
+  [Symbol.asyncIterator](): AsyncIterator<UdpDatagram>;
+};
 
-export type WasmInstance = {
-  exports: WasmExports;
+export type TcpListenerOptions = {
+  host?: string;
+  port: number;
+};
+
+export type TcpConnectionOptions = {
+  host: string;
+  port: number;
+};
+
+export type TcpConnection = {
+  readable: ReadableStream<Uint8Array>;
+  writable: WritableStream<Uint8Array>;
+  close(): Promise<void>;
+  [Symbol.asyncIterator](): AsyncIterator<Uint8Array>;
+};
+
+export type TcpListener = {
+  [Symbol.asyncIterator](): AsyncIterableIterator<TcpConnection>;
+};
+
+export type LoopbackInterfaceOptions = {
+  ip?: IPv4Cidr;
+};
+
+export type LoopbackInterface = {
+  readonly type: 'loopback';
+  readonly ip?: IPv4Address;
+  readonly netmask?: IPv4Address;
+};
+
+export type TunInterfaceOptions = {
+  ip?: IPv4Cidr;
+};
+
+export type TunInterface = {
+  readonly type: 'tun';
+  readonly ip?: IPv4Address;
+  readonly netmask?: IPv4Address;
+  readable: ReadableStream<Uint8Array>;
+  writable: WritableStream<Uint8Array>;
+  listen(): AsyncIterableIterator<Uint8Array>;
+  [Symbol.asyncIterator](): AsyncIterableIterator<Uint8Array>;
+};
+
+export type TapInterfaceOptions = {
+  mac?: MacAddress;
+  ip?: IPv4Cidr;
+};
+
+export type TapInterface = {
+  readonly type: 'tap';
+  readonly mac: MacAddress;
+  readonly ip?: IPv4Address;
+  readonly netmask?: IPv4Address;
+  readable: ReadableStream<Uint8Array>;
+  writable: WritableStream<Uint8Array>;
+  listen(): AsyncIterableIterator<Uint8Array>;
+  [Symbol.asyncIterator](): AsyncIterableIterator<Uint8Array>;
+};
+
+export type BridgeInterfaceOptions = {
+  ports: TapInterface[];
+  mac?: MacAddress;
+  ip?: IPv4Cidr;
+};
+
+export type BridgeInterface = {
+  readonly type: 'bridge';
+  readonly mac: MacAddress;
+  readonly ip?: IPv4Address;
+  readonly netmask?: IPv4Address;
 };
 
 export type NetworkInterface =
@@ -51,7 +113,34 @@ export type NetworkInterface =
   | TapInterface
   | BridgeInterface;
 
-export type DuplexStream<R = unknown> = {
-  readable: ReadableStream<R>;
-  writable: WritableStream<R>;
+export type NetworkStack = {
+  readonly ready: Promise<void>;
+  readonly interfaces: Iterable<NetworkInterface>;
+
+  createLoopbackInterface(
+    options: LoopbackInterfaceOptions
+  ): Promise<LoopbackInterface>;
+  createTunInterface(options: TunInterfaceOptions): Promise<TunInterface>;
+  createTapInterface(options?: TapInterfaceOptions): Promise<TapInterface>;
+  createBridgeInterface(
+    options: BridgeInterfaceOptions
+  ): Promise<BridgeInterface>;
+  removeInterface(
+    netInterface: LoopbackInterface | TunInterface | TapInterface
+  ): Promise<void>;
+  /**
+   * Listens for incoming TCP connections on the specified host/port.
+   */
+  listenTcp(options: TcpListenerOptions): Promise<TcpListener>;
+  /**
+   * Establishes an outbound TCP connection to a remote host/port.
+   */
+  connectTcp(options: TcpConnectionOptions): Promise<TcpConnection>;
+  /**
+   * Opens a UDP socket for sending and receiving datagrams.
+   *
+   * If no local host is provided, the socket will bind to all available interfaces.
+   * If no local port is provided, the socket will bind to a random port.
+   */
+  openUdp(options?: UdpSocketOptions): Promise<UdpSocket>;
 };
