+++ packages/tcpip/tsup.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from 'tsup';
 
 export default defineConfig([
   {
-    entry: ['src/index.ts'],
+    entry: ['src/index.ts', 'src/types.ts'],
     format: ['cjs', 'esm'],
     outDir: 'dist',
     sourcemap: true,
