+++ packages/v86/package.json
@@ -16,8 +16,8 @@
   "files": ["dist/**/*"],
   "exports": {
     ".": {
-      "import": "./dist/index.js",
       "types": "./dist/index.d.ts",
+      "import": "./dist/index.js",
       "default": "./dist/index.cjs"
     }
   },
