+++ packages/v86/src/network-adapter.test.ts
@@ -1,7 +1,7 @@
 import { type TcpSegment, parseEthernetFrame } from '@tcpip/wire';
 import { connectStreams, createStack } from 'tcpip';
 import { describe, expect, it } from 'vitest';
-import { createVm, nextValue } from '../test/util.ts';
+import { createVm, nextValue } from '../test/util.js';
 
 describe('network adapter', () => {
   it('should make tcp connection from VM to host', async () => {
