+++ packages/v86/tsconfig.json
@@ -1,5 +1,12 @@
 {
   "extends": "@total-typescript/tsconfig/tsc/dom/library",
   "include": ["src/**/*.ts"],
-  "exclude": ["src/**/*.test.ts"]
+  "compilerOptions": {
+    "paths": {
+      "tcpip": ["../tcpip/src/index.js"],
+      "tcpip/types": ["../tcpip/src/types.js"],
+      "@tcpip/dns": ["../dns/src/index.js"],
+      "@tcpip/wire": ["../wire/src/index.js"]
+    }
+  }
 }
