+++ packages/wire/package.json
@@ -14,8 +14,8 @@
   "files": ["dist/**/*"],
   "exports": {
     ".": {
-      "import": "./dist/index.js",
       "types": "./dist/index.d.ts",
+      "import": "./dist/index.js",
       "default": "./dist/index.cjs"
     }
   },
